requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-filter==24.2
pillow==10.3.0
openpyxl==3.1.5
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
gunicorn==23.0.0
drf-yasg==1.21.8
psycopg2==2.9.10
dj-database-url==2.3.0
